WebCL-conformance/conformance/bindingTesting/queues/cl_commandQueue_enqueueCopyImageToBuffer.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueCopyImageToBuffer conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<div style="position:absolute; left:0px; top:0px; visibility:hidden">
    <canvas id="dstBufferCanvas">
</div>
<script>
try {
    debug("WebCLCommandQueue :: enqueueCopyImageToBuffer() test case.");
    debug("( WebCLImage srcImage, WebCLBuffer dstBuffer, sequence<CLuint> srcOrigin, sequence<CLuint> srcRegion, CLuint dstOffset, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var origin = [0, 0];
    // width & height are chosen to match size of buffer.
    var width = 64;
    var height = 16;
    var region = [width, height];
    var dstOffset = 0;

    var webCLImageDescriptorReadWrite = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, width, height);
    var bufferSize = width * height * wtu.getBytesForChannelOrder(webCLImageDescriptorReadWrite.channelOrder) *
            eval(wtu.getArrayTypeForChanneltype(webCLImageDescriptorReadWrite.channelType) + ".BYTES_PER_ELEMENT");
    var webCLImageRead = wtu.createImage(webCLContext, webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite);
    var webCLImageReadWrite = wtu.createImage(webCLContext, webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
    var webCLImageWrite = wtu.createImage(webCLContext, webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);

    // Testing for MEM_WRITE_ONLY Buffer.
    var webCLBufferWrite = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, bufferSize);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageRead, webCLBufferWrite, origin, region, dstOffset);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferWrite, origin, region, dstOffset);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageWrite, webCLBufferWrite, origin, region, dstOffset);");

    // Testing for MEM_READ_WRITE Buffer.
    var webCLBufferReadWrite = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSize);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageRead, webCLBufferReadWrite, origin, region, dstOffset);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageWrite, webCLBufferReadWrite, origin, region, dstOffset);");

    // Testing for MEM_READ_ONLY Buffer.
    var webCLBufferRead = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, bufferSize);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageRead, webCLBufferRead, origin, region, dstOffset);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferRead, origin, region, dstOffset);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageWrite, webCLBufferRead, origin, region, dstOffset);");

    // Testing for nullable parameters.
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, []);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [], null);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, null);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, null, null);");
    var webCLEvent = wtu.createEvent();
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, null, webCLEvent);");

    // Testing with events.
    webCLEvent = wtu.createEvent();
    wtu.enqueueCopyImageToBuffer(webCLCommandQueue, webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [], webCLEvent);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [webCLEvent]);");
    var webCLUserEvent = wtu.createUserEvent(webCLContext);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [webCLUserEvent]);");
    wtu.setStatus(webCLUserEvent, webcl.COMPLETE);

    webCLEvent = wtu.createEvent();
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [], webCLEvent);");

    // Testing for exceptions.
    var releasedEvent = wtu.createUserEvent(webCLContext);
    wtu.release(releasedEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [releasedEvent], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

    webCLEvent = wtu.createEvent();
    wtu.enqueueCopyImageToBuffer(webCLCommandQueue, webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [webCLUserEvent], webCLEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffset, [webCLUserEvent], webCLEvent);", "INVALID_EVENT");

    // Testing for Exceptions.
    var releasedImage = wtu.createImage(webCLContext, webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);
    wtu.release(releasedImage);
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(releasedImage, webCLBufferReadWrite, origin, region, dstOffset);", "INVALID_MEM_OBJECT");
    var releasedBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, bufferSize);
    wtu.release(releasedBuffer);
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, releasedBuffer, origin, region, dstOffset);", "INVALID_MEM_OBJECT");

    // Testing for values of origin.
    var originInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, originInvalid, region, dstOffset);", "INVALID_VALUE");
    originInvalid = [width + 1, height + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, originInvalid, region, dstOffset);", "INVALID_VALUE");
    originInvalid = [0];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, originInvalid, region, dstOffset);", "INVALID_VALUE");

    // Testing for values of region.
    var regionInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, regionInvalid, dstOffset);", "INVALID_VALUE");
    regionInvalid = [width + 1, height + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, regionInvalid, dstOffset);", "INVALID_VALUE");
    regionInvalid = [width];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, regionInvalid, dstOffset);", "INVALID_VALUE");

    // Testing for values of destination Offset.
    var dstOffsetInvalid = bufferSize;
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, dstOffsetInvalid);", "INVALID_VALUE");

    // Testing for exception "INVALID_CONTEXT"
    var webCLContext1 = wtu.createContext();
    var webCLCommandQueue1 = wtu.createCommandQueue(webCLContext1);
    var webCLBuffer1 = wtu.createBuffer(webCLContext1, webcl.MEM_READ_WRITE, bufferSize);
    var webCLImageDescriptor1 = wtu.getSupportedImageFormats(webCLContext1, webcl.MEM_READ_WRITE, width, height);
    var webCLImage1 = wtu.createImage(webCLContext1, webcl.MEM_READ_WRITE, webCLImageDescriptor1);

    var webCLContext2 = wtu.createContext();
    var webCLCommandQueue2 = wtu.createCommandQueue(webCLContext2);
    var webCLBuffer2 = wtu.createBuffer(webCLContext2, webcl.MEM_READ_WRITE, bufferSize);
    var webCLImageDescriptor2 = wtu.getSupportedImageFormats(webCLContext2, webcl.MEM_READ_WRITE, width, height);
    var webCLImage2 = wtu.createImage(webCLContext2, webcl.MEM_READ_WRITE, webCLImageDescriptor2);

    shouldThrowExceptionName("webCLCommandQueue1.enqueueCopyImageToBuffer(webCLImage1, webCLBuffer2, origin, region, dstOffset)", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue1.enqueueCopyImageToBuffer(webCLImage2, webCLBuffer1, origin, region, dstOffset)", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue1.enqueueCopyImageToBuffer(webCLImage2, webCLBuffer2, origin, region, dstOffset)", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue2.enqueueCopyImageToBuffer(webCLImage2, webCLBuffer1, origin, region, dstOffset)", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue2.enqueueCopyImageToBuffer(webCLImage1, webCLBuffer2, origin, region, dstOffset)", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue2.enqueueCopyImageToBuffer(webCLImage1, webCLBuffer1, origin, region, dstOffset)", "INVALID_CONTEXT");

    var webCLUserEvent1 = wtu.createUserEvent(webCLContext1);
    var webCLUserEvent2 = wtu.createUserEvent(webCLContext2);
    shouldThrowExceptionName("webCLCommandQueue1.enqueueCopyImageToBuffer(webCLImage1, webCLBuffer1, origin, region, dstOffset, [webCLUserEvent2])", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue2.enqueueCopyImageToBuffer(webCLImage2, webCLBuffer2, origin, region, dstOffset, [webCLUserEvent1])", "INVALID_CONTEXT");
    wtu.setStatus(webCLUserEvent1, webcl.COMPLETE);
    wtu.setStatus(webCLUserEvent2, webcl.COMPLETE);

    var webCLEvent1 = wtu.createEvent();
    wtu.enqueueCopyImageToBuffer(webCLCommandQueue1, webCLImage1, webCLBuffer1, origin, region, dstOffset, [], webCLEvent1);
    shouldThrowExceptionName("webCLCommandQueue2.enqueueCopyImageToBuffer(webCLImage2, webCLBuffer2, origin, region, dstOffset, [webCLEvent1])", "INVALID_CONTEXT");
    var webCLEvent2 = wtu.createEvent();
    wtu.enqueueCopyImageToBuffer(webCLCommandQueue2, webCLImage2, webCLBuffer2, origin, region, dstOffset, [], webCLEvent2);
    shouldThrowExceptionName("webCLCommandQueue1.enqueueCopyImageToBuffer(webCLImage1, webCLBuffer1, origin, region, dstOffset, [webCLEvent2])", "INVALID_CONTEXT");

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
